create action to send a email to reset password
